Smart Class is added
