+++ src/views/smart.js
@@ -1,15 +1,42 @@
-import AbstractView from './abstract';
+import Abstract from './abstract.js';
 
-export default class Smart extends AbstractView {
-  restoreHandlers() {
-    alert('абстрактный метод');
+export default class Smart extends Abstract {
+  constructor() {
+    super();
+    this._data = {};
+  }
+
+  updateData(update, justDataUpdating) {
+    if (!update) {
+      return;
+    }
+
+    this._data = Object.assign(
+      {},
+      this._data,
+      update,
+    );
+
+    if (justDataUpdating) {
+      return;
+    }
+
+    this.updateElement();
   }
 
   updateElement() {
-    alert('обычный метод');
+    const prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    this.removeElement();
+
+    const newElement = this.getElement();
+
+    parent.replaceChild(newElement, prevElement);
+
+    this.restoreHandlers();
   }
 
-  updateData() {
-    alert('обычный метод');
+  restoreHandlers() {
+    throw new Error('Abstract method not implemented: resetHandlers');
   }
 }
